+++ frontend/src/drop_area.ts
@@ -128,7 +128,7 @@ class DropArea {
           const invalidFiles: File[] = [];
 
           for (const file of files) {
-            if (this.acceptedFileTypes.isAccepted(file.name)) {
+            if (this.acceptedFileTypes.isAccepted(file)) {
               acceptedFiles.push(file);
             } else {
               invalidFiles.push(file);